# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Review {
  id: number;
  name: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  reviewText: string;
  photoPath?: string | null;
  status: ReviewStatus;
  createdAt: string;
}

export interface SubmitReviewBody {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  reviewText: string;
  photoPath?: string | null;
}

export type AdminUpdateReviewBodyStatus = typeof AdminUpdateReviewBodyStatus[keyof typeof AdminUpdateReviewBodyStatus];


export const AdminUpdateReviewBodyStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface AdminUpdateReviewBody {
  status: AdminUpdateReviewBodyStatus;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorResponse {
  error: string;
}

export type AdminListReviewsParams = {
status?: AdminListReviewsStatus;
};

export type AdminListReviewsStatus = typeof AdminListReviewsStatus[keyof typeof AdminListReviewsStatus];


export const AdminListReviewsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;