# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reviews
    description: Customer review operations
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reviews:
    get:
      operationId: listApprovedReviews
      tags: [reviews]
      summary: List approved reviews
      description: Returns all approved customer reviews
      responses:
        "200":
          description: List of approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a review
      description: Submit a new customer review (stored as pending)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewBody"
      responses:
        "201":
          description: Review submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reviews:
    get:
      operationId: adminListReviews
      tags: [reviews]
      summary: Admin list all reviews
      description: Returns all reviews regardless of status (admin only)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
          required: false
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reviews/{id}:
    patch:
      operationId: adminUpdateReview
      tags: [reviews]
      summary: Admin approve or reject a review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateReviewBody"
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteReview
      tags: [reviews]
      summary: Admin delete a review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Review deleted
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Review:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
        photoPath:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - rating
        - reviewText
        - status
        - createdAt

    SubmitReviewBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
          minLength: 10
          maxLength: 2000
        photoPath:
          type: string
          nullable: true
      required:
        - name
        - rating
        - reviewText

    AdminUpdateReviewBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error