# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Returns all approved customer reviews
 * @summary List approved reviews
 */
export const listApprovedReviewsResponseRatingMax = 5;



export const ListApprovedReviewsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "rating": zod.number().min(1).max(listApprovedReviewsResponseRatingMax),
  "reviewText": zod.string(),
  "photoPath": zod.string().nullish(),
  "status": zod.enum(['pending', 'approved', 'rejected']),
  "createdAt": zod.coerce.date()
})
export const ListApprovedReviewsResponse = zod.array(ListApprovedReviewsResponseItem)


/**
 * Submit a new customer review (stored as pending)
 * @summary Submit a review
 */
export const submitReviewBodyNameMax = 100;

export const submitReviewBodyRatingMax = 5;

export const submitReviewBodyReviewTextMin = 10;
export const submitReviewBodyReviewTextMax = 2000;



export const SubmitReviewBody = zod.object({
  "name": zod.string().min(1).max(submitReviewBodyNameMax),
  "rating": zod.number().min(1).max(submitReviewBodyRatingMax),
  "reviewText": zod.string().min(submitReviewBodyReviewTextMin).max(submitReviewBodyReviewTextMax),
  "photoPath": zod.string().nullish()
})


/**
 * Returns all reviews regardless of status (admin only)
 * @summary Admin list all reviews
 */
export const AdminListReviewsQueryParams = zod.object({
  "status": zod.enum(['pending', 'approved', 'rejected']).optional()
})

export const adminListReviewsResponseRatingMax = 5;



export const AdminListReviewsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "rating": zod.number().min(1).max(adminListReviewsResponseRatingMax),
  "reviewText": zod.string(),
  "photoPath": zod.string().nullish(),
  "status": zod.enum(['pending', 'approved', 'rejected']),
  "createdAt": zod.coerce.date()
})
export const AdminListReviewsResponse = zod.array(AdminListReviewsResponseItem)


/**
 * @summary Admin approve or reject a review
 */
export const AdminUpdateReviewParams = zod.object({
  "id": zod.coerce.number()
})

export const AdminUpdateReviewBody = zod.object({
  "status": zod.enum(['approved', 'rejected'])
})

export const adminUpdateReviewResponseRatingMax = 5;



export const AdminUpdateReviewResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "rating": zod.number().min(1).max(adminUpdateReviewResponseRatingMax),
  "reviewText": zod.string(),
  "photoPath": zod.string().nullish(),
  "status": zod.enum(['pending', 'approved', 'rejected']),
  "createdAt": zod.coerce.date()
})


/**
 * @summary Admin delete a review
 */
export const AdminDeleteReviewParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string()
})